correción dibujo en el canvas

+++ src/main/java/edu/eci/arsw/magicBrushstrokesApp/model/Point.java
@@ -1,29 +1,29 @@
 package edu.eci.arsw.magicBrushstrokesApp.model;
 
 public class Point {
-    int x,y;
+    float x,y;
 
     public Point() {
     }
 
-    public Point(int x, int y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
